move hpy/microbench to the toplevel directory, the same as test/. We are not going to package it anyway
